# Reuse components slice when parsing chunk time range value.