@cyriltovena 2020-11-18
> LGTM Thanks.
> 
> This has a potential to introduce subtle bugs, if we leak `components` slice from any of the functions using the `componentsPool`. I've checked that current functions don't do that (conversions to `string` or `model.LabelValue` make a copy, and are safe). One more check to make sure we don't miss anything would be reassuring.

I plan to change the slice of string returned by a `[][]byte` in the future to avoid this said copy, because we mostly need the list of ids to fetch either chunks or other ids, the list of ids is never returned to the user. (aka it's only used for read).

As an example:

```
		chunkIDs, err := c.parseIndexEntries(ctx, entries, nil)
		if err != nil {
			return nil, err
		}
		totalChunkIDs = len(chunkIDs)

		return c.convertChunkIDsToChunks(ctx, userID, chunkIDs)
```

here the list of chunkIDs is again reslice and copied in the convertChunkIDsToChunks.

TLDR; in the future I want parseIndexEntries to return `[][]byte` and not `[]string` to avoid more unnecessary allocations.