@pstibrany 2020-11-18
Ah, I see. It's actually `components` slice itself that shouldn't leak, but individual elements are fine.